refactor: relocate sqlite-server components to data-space directory and update worker references
